+++ com/mattiacabrini/javads/Map.java
@@ -1,4 +1,4 @@
-package src;
+package com.mattiacabrini.javads;
 
 import java.util.ArrayList;
 import java.util.Iterator;
